move heroicon to sharedComponents
